+++ lang/en/edusharing.php
@@ -155,8 +155,8 @@ $string['obfuscate_auth_param'] = 'User ID pseudonymization';
 $string['obfuscate_auth_param_help'] = 'If activated, Moodle users will be pseudonymized in the Edu-Sharing repository';
 $string['require_login_for_metadata'] = 'Login required for metadata';
 $string['require_login_for_metadata_help'] = 'If activated, app metadata can only be queried after login';
-
-$string['soaprequired'] = 'The PHP extension soap must be activated.';
+$string['use_as_idp'] = 'Use Moodle as IDP for Edu-Sharing';
+$string['use_as_idp_help'] = 'If activated, Moodle can be used as an IDP for Edu-Sharing login.';
 
 $string['error_missing_authwsdl'] = 'No "authenticationwebservice_wsdl" configured.';
 $string['error_authservice_not_reachable'] = 'not reachable. Cannot utilize edu-sharing network.';